Fixed integer overflow in Device::getTextureDataForText (in CCDevice-ios.mm)
There was the problem when “height” and “width” were typecast to short
integer. So for values more than sizeof(short) we had integer overflow.

Steps to reproduce:
Create simple application with EditBox. Start it. Add 1000 (just
example) characters at a time to the EditBox. Method
Device::getTextureDataForText will be called and there will be integer
overflow for “height” and “width”. As a result application will crash
later.

+++ cocos/platform/ios/CCDevice-ios.mm
@@ -485,8 +485,8 @@ Data Device::getTextureDataForText(const char * text, const FontDefinition& text
         {
             break;
         }
-        height = (short)info.height;
-        width = (short)info.width;
+        height = info.height;
+        width = info.width;
         ret.fastSet(info.data,width * height * 4);
         hasPremultipliedAlpha = true;
     } while (0);
